Rename cv2 stub and patch tests
