+++ cv2_stub.py
